refactor!(test, db): move unit trait definitions into test module
As a side-effect, all integrations now accept a tempdir.
