+++ src/db/traits.rs
@@ -1,12 +1,11 @@
 use std::fmt::Debug;
 use std::ops::RangeBounds;
-use std::{collections::BTreeMap, convert::Infallible, fmt::Display};
+use std::{collections::BTreeMap, fmt::Display};
 
 use bitcoin::{block::Header, BlockHash};
 
 use crate::prelude::FutureResult;
 
-use super::error::UnitPeerStoreError;
 use super::PersistedPeer;
 
 /// Methods required to persist the chain of block headers.
@@ -45,65 +44,6 @@ pub trait HeaderStore: Debug + Send + Sync {
     fn header_at(&mut self, height: u32) -> FutureResult<Option<Header>, Self::Error>;
 }
 
-/// This is a simple wrapper for the unit type, signifying that no headers will be stored between sessions.
-impl HeaderStore for () {
-    type Error = Infallible;
-    fn load<'a>(
-        &'a mut self,
-        _range: impl RangeBounds<u32> + Send + Sync + 'a,
-    ) -> FutureResult<'a, BTreeMap<u32, Header>, Self::Error> {
-        async fn do_load() -> Result<BTreeMap<u32, Header>, Infallible> {
-            Ok(BTreeMap::new())
-        }
-        Box::pin(do_load())
-    }
-
-    fn write<'a>(
-        &'a mut self,
-        _header_chain: &'a BTreeMap<u32, Header>,
-    ) -> FutureResult<'a, (), Self::Error> {
-        async fn do_write() -> Result<(), Infallible> {
-            Ok(())
-        }
-        Box::pin(do_write())
-    }
-
-    fn write_over<'a>(
-        &'a mut self,
-        _header_chain: &'a BTreeMap<u32, Header>,
-        _height: u32,
-    ) -> FutureResult<'a, (), Self::Error> {
-        async fn do_write_over() -> Result<(), Infallible> {
-            Ok(())
-        }
-        Box::pin(do_write_over())
-    }
-
-    fn height_of<'a>(
-        &'a mut self,
-        _block_hash: &'a BlockHash,
-    ) -> FutureResult<'a, Option<u32>, Self::Error> {
-        async fn do_height_of() -> Result<Option<u32>, Infallible> {
-            Ok(None)
-        }
-        Box::pin(do_height_of())
-    }
-
-    fn hash_at(&mut self, _height: u32) -> FutureResult<Option<BlockHash>, Self::Error> {
-        async fn do_hast_at() -> Result<Option<BlockHash>, Infallible> {
-            Ok(None)
-        }
-        Box::pin(do_hast_at())
-    }
-
-    fn header_at(&mut self, _height: u32) -> FutureResult<Option<Header>, Self::Error> {
-        async fn do_header_at() -> Result<Option<Header>, Infallible> {
-            Ok(None)
-        }
-        Box::pin(do_header_at())
-    }
-}
-
 /// Methods that define a list of peers on the Bitcoin P2P network.
 pub trait PeerStore: Debug + Send + Sync {
     /// Errors that may occur within a [`PeerStore`].
@@ -118,26 +58,105 @@ pub trait PeerStore: Debug + Send + Sync {
     fn num_unbanned(&mut self) -> FutureResult<u32, Self::Error>;
 }
 
-impl PeerStore for () {
-    type Error = UnitPeerStoreError;
-    fn update(&mut self, _peer: PersistedPeer) -> FutureResult<(), Self::Error> {
-        async fn do_update() -> Result<(), UnitPeerStoreError> {
-            Ok(())
-        }
-        Box::pin(do_update())
+#[cfg(test)]
+mod test {
+    use super::*;
+    use std::convert::Infallible;
+
+    /// Errors for the [`PeerStore`](crate) of unit type.
+    #[derive(Debug)]
+    pub enum UnitPeerStoreError {
+        /// There were no peers found.
+        NoPeers,
     }
 
-    fn random(&mut self) -> FutureResult<PersistedPeer, Self::Error> {
-        async fn do_random() -> Result<PersistedPeer, UnitPeerStoreError> {
-            Err(UnitPeerStoreError::NoPeers)
+    impl core::fmt::Display for UnitPeerStoreError {
+        fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+            match self {
+                UnitPeerStoreError::NoPeers => write!(f, "no peers in unit database."),
+            }
         }
-        Box::pin(do_random())
     }
 
-    fn num_unbanned(&mut self) -> FutureResult<u32, Self::Error> {
-        async fn do_num_unbanned() -> Result<u32, UnitPeerStoreError> {
-            Ok(0)
+    impl PeerStore for () {
+        type Error = UnitPeerStoreError;
+        fn update(&mut self, _peer: PersistedPeer) -> FutureResult<(), Self::Error> {
+            async fn do_update() -> Result<(), UnitPeerStoreError> {
+                Ok(())
+            }
+            Box::pin(do_update())
         }
-        Box::pin(do_num_unbanned())
+
+        fn random(&mut self) -> FutureResult<PersistedPeer, Self::Error> {
+            async fn do_random() -> Result<PersistedPeer, UnitPeerStoreError> {
+                Err(UnitPeerStoreError::NoPeers)
+            }
+            Box::pin(do_random())
+        }
+
+        fn num_unbanned(&mut self) -> FutureResult<u32, Self::Error> {
+            async fn do_num_unbanned() -> Result<u32, UnitPeerStoreError> {
+                Ok(0)
+            }
+            Box::pin(do_num_unbanned())
+        }
+    }
+
+    impl HeaderStore for () {
+        type Error = Infallible;
+        fn load<'a>(
+            &'a mut self,
+            _range: impl RangeBounds<u32> + Send + Sync + 'a,
+        ) -> FutureResult<'a, BTreeMap<u32, Header>, Self::Error> {
+            async fn do_load() -> Result<BTreeMap<u32, Header>, Infallible> {
+                Ok(BTreeMap::new())
+            }
+            Box::pin(do_load())
+        }
+
+        fn write<'a>(
+            &'a mut self,
+            _header_chain: &'a BTreeMap<u32, Header>,
+        ) -> FutureResult<'a, (), Self::Error> {
+            async fn do_write() -> Result<(), Infallible> {
+                Ok(())
+            }
+            Box::pin(do_write())
+        }
+
+        fn write_over<'a>(
+            &'a mut self,
+            _header_chain: &'a BTreeMap<u32, Header>,
+            _height: u32,
+        ) -> FutureResult<'a, (), Self::Error> {
+            async fn do_write_over() -> Result<(), Infallible> {
+                Ok(())
+            }
+            Box::pin(do_write_over())
+        }
+
+        fn height_of<'a>(
+            &'a mut self,
+            _block_hash: &'a BlockHash,
+        ) -> FutureResult<'a, Option<u32>, Self::Error> {
+            async fn do_height_of() -> Result<Option<u32>, Infallible> {
+                Ok(None)
+            }
+            Box::pin(do_height_of())
+        }
+
+        fn hash_at(&mut self, _height: u32) -> FutureResult<Option<BlockHash>, Self::Error> {
+            async fn do_hast_at() -> Result<Option<BlockHash>, Infallible> {
+                Ok(None)
+            }
+            Box::pin(do_hast_at())
+        }
+
+        fn header_at(&mut self, _height: u32) -> FutureResult<Option<Header>, Self::Error> {
+            async fn do_header_at() -> Result<Option<Header>, Infallible> {
+                Ok(None)
+            }
+            Box::pin(do_header_at())
+        }
     }
 }
